pyproject.toml
[project]
name = "brainstorm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
